ui: 습관 보관함 삭제하기 ui 적용
# Conflicts:
#	data/src/main/java/com/depromeet/threedays/data/datasource/habit/HabitRemoteDataSourceImpl.kt
#	presentation/mypage/src/main/res/values/strings.xml

+++ data/src/main/java/com/depromeet/threedays/data/entity/HabitEntity.kt
@@ -3,7 +3,7 @@ package com.depromeet.threedays.data.entity
 import com.depromeet.threedays.domain.entity.mate.Mate
 
 class HabitEntity (
-    val habitId: Int,
+    val habitId: Long,
     val memberId: Int,
     val title: String,
     val imojiPath: String,
@@ -11,7 +11,7 @@ class HabitEntity (
     val reward: Int,
     val color: String,
     val mate: Mate?,
-    val todayHabitAchievementId: Int?, // null이면 오늘 체크를 하지 않은 것
+    val todayHabitAchievementId: Long?, // null이면 오늘 체크를 하지 않은 것
     val sequence: Int,
     val createAt: String,
 )
