+++ domain/src/main/java/com/depromeet/threedays/domain/entity/habit/Habit.kt
@@ -4,7 +4,7 @@ import com.depromeet.threedays.domain.entity.Color
 import com.depromeet.threedays.domain.entity.mate.Mate
 
 data class Habit(
-    val habitId: Int,
+    val habitId: Long,
     val memberId: Int,
     val title: String,
     val imojiPath: String,
@@ -12,7 +12,7 @@ data class Habit(
     val reward: Int,
     val color: Color,
     val mate: Mate?,
-    val todayHabitAchievementId: Int?, // null이면 오늘 체크를 하지 않은 것
+    val todayHabitAchievementId: Long?, // null이면 오늘 체크를 하지 않은 것
     val sequence: Int,
     val createAt: String,
 )
